backend/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Daniil Bezuglov", email = "daniil.bezuglov@lockerbot.io"},
]
dependencies = [
    "django>=4.1.2",
    "djangorestframework>=3.14.0",
    "django-filter>=22.1",
    "pip>=22.3",
    "django-model-utils>=4.2.0",
    "setuptools>=65.5.0",
    "djangorestframework-api-key>=2.2.0",
    "djangorestframework-simplejwt>=5.2.1",
    "regex>=2022.9.13",
    "drf-yasg>=1.21.4",
    "cryptography>=38.0.1",
    "django-cors-headers>=3.13.0",
]
requires-python = ">=3.10"
license = {text = "None"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]

[tool.pdm]

[tool.pdm.scripts]
start = "python manage.py runserver"

